fix: update schema and resolve issues.

+++ tutionmanagement/src/main/java/com/safwan/tutionmanagement/service/ReportService.java
@@ -1,10 +1,13 @@
 package com.safwan.tutionmanagement.service;
 
+import com.safwan.tutionmanagement.modal.Attendance;
 import com.safwan.tutionmanagement.modal.Report;
+import com.safwan.tutionmanagement.modal.ReportMarks;
 import com.safwan.tutionmanagement.repo.ReportRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,6 +18,11 @@ public class ReportService {
     private ReportRepository reportRepository;
 
     public Report saveReport(Report report) {
+        if (report.getReportMarks() != null) {
+            for (ReportMarks mark : report.getReportMarks()) {
+                mark.setReportCard(report);
+            }
+        }
         return reportRepository.save(report);
     }
 
@@ -28,6 +36,33 @@ public class ReportService {
     }
 
     public void deleteReport(Long id) {
-        reportRepository.deleteById(id);
+        Report report = getReportById(id);
+
+        if(report != null) {
+            reportRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Report not found while deleting it, id: " + id);
+        }
+    }
+
+    public Report updateReport(Long id, Report newReport) {
+        Report existingRep = getReportById(id);
+
+        //clear old children
+        existingRep.getReportMarks().clear();
+
+        //Add new Children
+        for (ReportMarks mark : newReport.getReportMarks()){
+            mark.setReportCard(existingRep);
+            existingRep.getReportMarks().add(mark);
+        }
+
+        existingRep.setExamName(newReport.getExamName());
+        existingRep.setDate(newReport.getDate());
+        existingRep.setMonth(newReport.getMonth());
+        existingRep.setYear(newReport.getYear());
+        existingRep.setStudent(newReport.getStudent());
+
+        return reportRepository.save(existingRep);
     }
 }
